Add tests and remove unnecessary Filter Vi
The Filter VI is not needed in the API or internally because the Alphabetize VI automatically filters for Clusters and ClusterConstants. Note, that Cluster controls appears as terminals on the block diagram and cannot be made more specific to clusters. You need to get GObjects for the Front Panel (FP) and Block Diagram (BD) separately and combine them if all of the clusters, both controls/indicators and constants, are to be alphabetized in a VI.

git-svn-id: svn://localhost/Alphabetize%20Cluster%20Elements/trunk@11 87d7cce1-06af-c242-8b3f-723c77aec3b0