Fix Alphabetize menu item not appearing for any clusters
When a cluster is right-clicked, either for a constant or control, multiple affected GObjects are included, not just the cluster. For example, right-clicking on a cluster constant will select two affected items: ClusterConstant and Terminal. So, counting the number of affected items when building the menu does not work. Instead, we just search for any cluster in the affected items and let the menu item appear only if the Autosizing menu item also appears.

git-svn-id: svn://localhost/Alphabetize%20Cluster%20Elements/trunk@31 87d7cce1-06af-c242-8b3f-723c77aec3b0